pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "autoee-demo"
version = "0.1.0"
description = "AutoEE Buck workflow demo with pluggable model backends."
requires-python = ">=3.10"
dependencies = [
  "PyQt5>=5.15",
  "matplotlib>=3.8",
  "requests>=2.31",
  "keyring>=25.0",
  "PyYAML>=6.0",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "pywebview>=5.0; python_version < '3.14'",
]

[project.optional-dependencies]
webview = [
  "pywebview>=5.0",
]

[tool.setuptools.packages.find]
include = ["autoee_demo*"]
